@@ -1,5 +1,3 @@
-import 'package:meta/meta.dart';
-
 abstract class SaveSecureCacheStorage {
-  Future<void> saveSecure({@required String key, @required String value});
+  Future<void> saveSecure({required String key, required String? value});
 }
