@@ -1,9 +1,7 @@
-import 'package:meta/meta.dart';
-
 abstract class HttpClient {
-  Future<Map> request({
-    @required String url,
-    @required String method,
-    Map body,
+  Future<Map?> request({
+    required String url,
+    required String method,
+    Map? body,
   });
 }
