@@ -3,7 +3,7 @@ import '../../domain/entities/entities.dart';
 import '../http/http.dart';
 
 class RemoteAccountModel {
-  final String accessToken;
+  final String? accessToken;
 
   RemoteAccountModel(this.accessToken);
 
