@@ -1,5 +1,3 @@
-import 'package:meta/meta.dart';
-
 import '../../../domain/entities/entities.dart';
 import '../../../domain/helpers/helpers.dart';
 import '../../../domain/usecases/usecases.dart';
@@ -9,12 +7,11 @@ import '../../cache/cache.dart';
 class LocalSaveCurrentAccout implements SaveCurrentAccount {
   final SaveSecureCacheStorage saveSecureCacheStorage;
 
-  LocalSaveCurrentAccout({@required this.saveSecureCacheStorage});
+  LocalSaveCurrentAccout({required this.saveSecureCacheStorage});
 
   Future<void> save(AccountEntity account) async {
     try {
-      await saveSecureCacheStorage.saveSecure(
-          key: 'token', value: account.token);
+      await saveSecureCacheStorage.saveSecure(key: 'token', value: account.token);
     } catch (error) {
       throw DomainError.unexpected;
     }
