@@ -1,7 +1,7 @@
 import 'package:equatable/equatable.dart';
 
 class AccountEntity extends Equatable {
-  final String token;
+  final String? token;
 
   @override
   List get props => [token];
