@@ -1,5 +1,4 @@
 import 'package:equatable/equatable.dart';
-import 'package:meta/meta.dart';
 
 import '../entities/entities.dart';
 
@@ -8,14 +7,14 @@ abstract class Authentication {
 }
 
 class AuthenticationParams extends Equatable {
-  final String email;
-  final String password;
+  final String? email;
+  final String? password;
 
   @override
   List get props => [email, password];
 
   AuthenticationParams({
-    @required this.email,
-    @required this.password,
+    required this.email,
+    required this.password,
   });
 }
