@@ -1,20 +1,17 @@
 import 'package:flutter_secure_storage/flutter_secure_storage.dart';
-import 'package:meta/meta.dart';
 
 import '../../data/cache/cache.dart';
 
-class LocalStorageAdapter
-    implements SaveSecureCacheStorage, FetchSecureCacheStorage {
+class LocalStorageAdapter implements SaveSecureCacheStorage, FetchSecureCacheStorage {
   final FlutterSecureStorage secureStorage;
 
-  LocalStorageAdapter({@required this.secureStorage});
+  LocalStorageAdapter({required this.secureStorage});
 
-  Future<void> saveSecure(
-      {@required String key, @required String value}) async {
+  Future<void> saveSecure({required String key, required String? value}) async {
     await secureStorage.write(key: key, value: value);
   }
 
-  Future<String> fetchSecure(String key) async {
+  Future<String?> fetchSecure(String key) async {
     final value = await secureStorage.read(key: key);
     return value;
   }
