@@ -1,18 +1,17 @@
 import 'dart:convert';
 
 import 'package:http/http.dart';
-import 'package:meta/meta.dart';
 
 import '../../data/http/http.dart';
 
 class HttpAdapter implements HttpClient {
-  final Client client;
+  final Client? client;
   HttpAdapter(this.client);
 
-  Future<Map> request({
-    @required String url,
-    @required String method,
-    Map body,
+  Future<Map?> request({
+    required String? url,
+    required String method,
+    Map? body,
   }) async {
     final headers = {
       'Content-Type': 'application/json',
@@ -22,8 +21,7 @@ class HttpAdapter implements HttpClient {
     var response = Response('', 500);
     try {
       if (method == 'post') {
-        response =
-            await client.post(Uri.parse(url), headers: headers, body: jsonBody);
+        response = await client!.post(Uri.parse(url!), headers: headers, body: jsonBody);
       }
     } catch (error) {
       throw HttpError.serverError;
@@ -31,7 +29,7 @@ class HttpAdapter implements HttpClient {
     return _handleResponse(response);
   }
 
-  Map _handleResponse(Response response) {
+  Map? _handleResponse(Response response) {
     if (response.statusCode == 200) {
       return response.body.isEmpty ? null : jsonDecode(response.body);
     } else if (response.statusCode == 204) {
