@@ -2,15 +2,15 @@ import '../../validation/protocols/protocols.dart';
 import '../../validation/validators/validators.dart';
 
 class ValidationBuilder {
-  static ValidationBuilder _instance;
-  String fieldName;
+  static ValidationBuilder? _instance;
+  String? fieldName;
   List<FieldValidation> validations = [];
 
   ValidationBuilder._();
 
-  static ValidationBuilder field(String fieldName) {
+  static ValidationBuilder? field(String fieldName) {
     _instance = ValidationBuilder._();
-    _instance.fieldName = fieldName;
+    _instance!.fieldName = fieldName;
     return _instance;
   }
 
