@@ -1,6 +1,5 @@
-import 'package:get/get_state_manager/get_state_manager.dart';
+import 'package:get/get.dart';
 import 'package:get/state_manager.dart';
-import 'package:meta/meta.dart';
 
 import '../../ui/pages/pages.dart';
 import '../../ui/helpers/helpers.dart';
@@ -11,46 +10,46 @@ import '../../domain/usecases/usecases.dart';
 import '../protocols/protocols.dart';
 
 class GetxLoginPresenter extends GetxController implements LoginPresenter {
-  final Validation validation;
-  final Authentication authentication;
-  final SaveCurrentAccount saveCurrentAccount;
+  final Validation? validation;
+  final Authentication? authentication;
+  final SaveCurrentAccount? saveCurrentAccount;
 
-  String _email;
-  String _password;
-  var _emailError = Rx<UIError>(null);
-  var _passwordError = Rx<UIError>(null);
-  var _mainError = Rx<UIError>(null);
-  var _navigateTo = RxString(null);
+  String? _email;
+  String? _password;
+  var _emailError = Rx<UIError?>(null);
+  var _passwordError = Rx<UIError?>(null);
+  var _mainError = Rx<UIError?>(null);
+  var _navigateTo = RxString('');
   var _isFormValid = false.obs;
   var _isLoading = false.obs;
 
-  Stream<UIError> get emailErrorStream => _emailError.stream;
-  Stream<UIError> get passwordErrorStream => _passwordError.stream;
-  Stream<UIError> get mainErrorStream => _mainError.stream;
+  Stream<UIError?> get emailErrorStream => _emailError.stream;
+  Stream<UIError?> get passwordErrorStream => _passwordError.stream;
+  Stream<UIError?> get mainErrorStream => _mainError.stream;
   Stream<String> get navigateToStream => _navigateTo.stream;
   Stream<bool> get isFormValidStream => _isFormValid.stream;
   Stream<bool> get isLoadingStream => _isLoading.stream;
 
   GetxLoginPresenter({
-    @required this.validation,
-    @required this.authentication,
-    @required this.saveCurrentAccount,
+    required this.validation,
+    required this.authentication,
+    required this.saveCurrentAccount,
   });
 
-  void validateEmail(String email) {
+  void validateEmail(String? email) {
     _email = email;
     _emailError.value = _validateField(field: 'email', value: email);
     _validateForm();
   }
 
-  void validatePassword(String password) {
+  void validatePassword(String? password) {
     _password = password;
     _passwordError.value = _validateField(field: 'password', value: password);
     _validateForm();
   }
 
-  UIError _validateField({String field, String value}) {
-    final error = validation.validate(field: field, value: value);
+  UIError? _validateField({String? field, String? value}) {
+    final error = validation!.validate(field: field, value: value);
     switch (error) {
       case ValidationError.invalidField:
         return UIError.invalidField;
@@ -72,8 +71,8 @@ class GetxLoginPresenter extends GetxController implements LoginPresenter {
     try {
       _isLoading.value = true;
       final account =
-          await authentication.auth(AuthenticationParams(email: _email, password: _password));
-      await saveCurrentAccount.save(account);
+          await authentication!.auth(AuthenticationParams(email: _email, password: _password));
+      await saveCurrentAccount!.save(account);
       _navigateTo.value = '/surveys';
     } on DomainError catch (error) {
       switch (error) {
