@@ -1,5 +1,4 @@
 import 'package:get/get.dart';
-import 'package:meta/meta.dart';
 
 import '../../domain/usecases/usecases.dart';
 import '../../ui/pages/pages.dart';
@@ -7,17 +6,17 @@ import '../../ui/pages/pages.dart';
 class GetxSplashPresenter extends SplashPresenter {
   LoadCurrentAccount loadCurrentAccount;
 
-  var _navigateTo = RxString(null);
+  var _navigateTo = RxString('');
 
   Stream<String> get navigateToStream => _navigateTo.stream;
 
-  GetxSplashPresenter({@required this.loadCurrentAccount});
+  GetxSplashPresenter({required this.loadCurrentAccount});
 
-  Future<void> checkAccount({int durationInSeconds = 3}) async {
-    await Future.delayed(Duration(seconds: durationInSeconds));
+  Future<void> checkAccount({int? durationInSeconds = 3}) async {
+    await Future.delayed(Duration(seconds: durationInSeconds!));
     try {
       final account = await loadCurrentAccount.load();
-      _navigateTo.value = account == null ? '/login' : '/surveys';
+      _navigateTo.value = account.token == '' ? '/login' : '/surveys';
     } catch (error) {
       _navigateTo.value = '/login';
     }
