@@ -1,7 +1,5 @@
-import 'package:meta/meta.dart';
-
 abstract class Validation {
-  ValidationError validate({@required String field, @required String value});
+  ValidationError? validate({required String? field, required String? value});
 }
 
 enum ValidationError {
