@@ -5,17 +5,17 @@ import 'package:get/get.dart';
 import './splash_presenter.dart';
 
 class SplashPage extends StatelessWidget {
-  final SplashPresenter presenter;
+  final SplashPresenter? presenter;
 
-  SplashPage({@required this.presenter});
+  SplashPage({required this.presenter});
 
   @override
   Widget build(BuildContext context) {
-    presenter.checkAccount();
+    presenter!.checkAccount();
     return Scaffold(
       body: Builder(builder: (context) {
-        presenter.navigateToStream.listen((page) {
-          if (page?.isNotEmpty == true) {
+        presenter!.navigateToStream.listen((page) {
+          if (page.isNotEmpty == true) {
             Get.offAllNamed(page);
           }
         });
