@@ -1,5 +1,3 @@
-import 'package:meta/meta.dart';
-
 import '../../presentation/protocols/protocols.dart';
 
 import '../protocols/protocols.dart';
@@ -9,10 +7,9 @@ class ValidationComposite implements Validation {
 
   ValidationComposite(this.validations);
 
-  ValidationError validate({@required String field, @required String value}) {
-    ValidationError error;
-    for (final validation
-        in validations.where((element) => element.field == field)) {
+  ValidationError? validate({required String? field, required String? value}) {
+    ValidationError? error;
+    for (final validation in validations.where((element) => element.field == field)) {
       error = validation.validate(value);
       if (error != null) {
         return error;
