@@ -1,4 +1,5 @@
 import 'package:faker/faker.dart';
+import 'package:mockito/annotations.dart';
 import 'package:mockito/mockito.dart';
 import 'package:test/test.dart';
 
@@ -8,16 +9,15 @@ import 'package:flutter_app/domain/entities/entities.dart';
 import 'package:flutter_app/data/cache/cache.dart';
 import 'package:flutter_app/data/usecases/usecases.dart';
 
-class FetchSecureCacheStorageSpy extends Mock
-    implements FetchSecureCacheStorage {}
+import 'local_load_current_account_test.mocks.dart';
 
+@GenerateMocks([FetchSecureCacheStorage])
 void main() {
-  LocalLoadCurrentAccount sut;
-  FetchSecureCacheStorageSpy fetchSecureCacheStorage;
-  String token;
+  late LocalLoadCurrentAccount sut;
+  late MockFetchSecureCacheStorage fetchSecureCacheStorage;
+  late String token;
 
-  PostExpectation mockFetchSecureCall() =>
-      when(fetchSecureCacheStorage.fetchSecure(any));
+  PostExpectation mockFetchSecureCall() => when(fetchSecureCacheStorage.fetchSecure(any));
 
   void mockFetchSecure() {
     mockFetchSecureCall().thenAnswer((_) async => token);
@@ -28,9 +28,8 @@ void main() {
   }
 
   setUp(() {
-    fetchSecureCacheStorage = FetchSecureCacheStorageSpy();
-    sut = LocalLoadCurrentAccount(
-        fetchSecureCacheStorage: fetchSecureCacheStorage);
+    fetchSecureCacheStorage = MockFetchSecureCacheStorage();
+    sut = LocalLoadCurrentAccount(fetchSecureCacheStorage: fetchSecureCacheStorage);
     token = faker.guid.guid();
     mockFetchSecure();
   });
@@ -47,8 +46,7 @@ void main() {
     expect(account, AccountEntity(token));
   });
 
-  test('Should throw UnexpectedError if FetchSecureCacheStorage throws',
-      () async {
+  test('Should throw UnexpectedError if FetchSecureCacheStorage throws', () async {
     mockFetchSecureError();
 
     final future = sut.load();
