@@ -1,4 +1,5 @@
 import 'package:faker/faker.dart';
+import 'package:mockito/annotations.dart';
 import 'package:mockito/mockito.dart';
 import 'package:test/test.dart';
 
@@ -8,35 +9,31 @@ import 'package:flutter_app/domain/entities/entities.dart';
 import 'package:flutter_app/data/cache/cache.dart';
 import 'package:flutter_app/data/usecases/usecases.dart';
 
-class SaveSecureCacheStorageSpy extends Mock implements SaveSecureCacheStorage {
-}
+import 'local_save_current_account_test.mocks.dart';
 
+@GenerateMocks([SaveSecureCacheStorage])
 void main() {
-  SaveSecureCacheStorageSpy saveSecureCacheStorage;
-  LocalSaveCurrentAccout sut;
-  AccountEntity account;
+  late MockSaveSecureCacheStorage saveSecureCacheStorage;
+  late LocalSaveCurrentAccout sut;
+  late AccountEntity account;
 
   setUp(() {
-    saveSecureCacheStorage = SaveSecureCacheStorageSpy();
-    sut =
-        LocalSaveCurrentAccout(saveSecureCacheStorage: saveSecureCacheStorage);
+    saveSecureCacheStorage = MockSaveSecureCacheStorage();
+    sut = LocalSaveCurrentAccout(saveSecureCacheStorage: saveSecureCacheStorage);
     account = AccountEntity(faker.guid.guid());
   });
 
   void mockError() {
-    when(saveSecureCacheStorage.saveSecure(
-            key: anyNamed('key'), value: anyNamed('value')))
+    when(saveSecureCacheStorage.saveSecure(key: anyNamed('key'), value: anyNamed('value')))
         .thenThrow(Exception());
   }
 
   test('Should call SaveSecureCacheStorage with correct values', () async {
     await sut.save(account);
-    verify(
-        saveSecureCacheStorage.saveSecure(key: 'token', value: account.token));
+    verify(saveSecureCacheStorage.saveSecure(key: 'token', value: account.token));
   });
 
-  test('Should throw UnexpectedError if SaveSecureCacheStorage throws',
-      () async {
+  test('Should throw UnexpectedError if SaveSecureCacheStorage throws', () async {
     mockError();
 
     final future = sut.save(account);
