@@ -9,7 +9,7 @@ class EmailInput extends StatelessWidget {
   @override
   Widget build(BuildContext context) {
     final presenter = Provider.of<LoginPresenter>(context);
-    return StreamBuilder<UIError>(
+    return StreamBuilder<UIError?>(
         stream: presenter.emailErrorStream,
         builder: (context, snapshot) {
           return TextFormField(
