@@ -8,7 +8,7 @@ class PasswordInput extends StatelessWidget {
   @override
   Widget build(BuildContext context) {
     final presenter = Provider.of<LoginPresenter>(context);
-    return StreamBuilder<UIError>(
+    return StreamBuilder<UIError?>(
         stream: presenter.passwordErrorStream,
         builder: (context, snapshot) {
           return TextFormField(
