@@ -1,10 +1,10 @@
 import '../../helpers/helpers.dart';
 
 abstract class LoginPresenter {
-  Stream<UIError> get emailErrorStream;
-  Stream<UIError> get passwordErrorStream;
-  Stream<UIError> get mainErrorStream;
-  Stream<String> get navigateToStream;
+  Stream<UIError?> get emailErrorStream;
+  Stream<UIError?> get passwordErrorStream;
+  Stream<UIError?> get mainErrorStream;
+  Stream<String?> get navigateToStream;
   Stream<bool> get isFormValidStream;
   Stream<bool> get isLoadingStream;
 
