@@ -1,5 +1,5 @@
 abstract class SplashPresenter {
   Stream<String> get navigateToStream;
 
-  Future<void> checkAccount({int durationInSeconds});
+  Future<void> checkAccount({int? durationInSeconds});
 }
