@@ -1,6 +1,6 @@
 import '../../presentation/protocols/protocols.dart';
 
 abstract class FieldValidation {
-  String get field;
-  ValidationError validate(String value);
+  String? get field;
+  ValidationError? validate(String? value);
 }
